externalplugin refactor
